Added & refactored flyout samples
