+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/Flyout_Events.xaml
