+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/Flyout_Simple.xaml
@@ -115,7 +115,7 @@
 			</Button.Flyout>
 		  </Button>
 
-		  <Button Content="Show Flyout With Offset"
+		  <Button Content="Show Flyout W/offset"
 				  Margin="100"
 				  Height="50"
 				  Width="150">
