+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/Flyout_Simple.xaml.cs
@@ -1,0 +1,14 @@
+using Uno.UI.Samples.Controls;
+using Windows.UI.Xaml.Controls;
+
+namespace Uno.UI.Samples.Content.UITests.Flyout
+{
+	[SampleControlInfoAttribute("Flyout", "Flyout_Simple")]
+	public sealed partial class Flyout_Simple : UserControl
+	{
+		public Flyout_Simple()
+		{
+			this.InitializeComponent();
+		}
+	}
+}
