+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/Flyout_Vanilla.xaml.cs
@@ -1,0 +1,23 @@
+﻿using Uno.UI.Samples.Controls;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+
+namespace UITests.Shared.Windows_UI_Xaml_Controls.FlyoutTests
+{
+	[SampleControlInfo("Flyout", "Vanilla")]
+	public sealed partial class Flyout_Vanilla : UserControl
+	{
+		public Flyout_Vanilla()
+		{
+			this.InitializeComponent();
+		}
+
+		private void DeleteConfirmation_Click(object sender, RoutedEventArgs e)
+		{
+			if(Control1.Flyout is Flyout f)
+			{
+				f.Hide();
+			}
+		}
+	}
+}
