+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/MenuFlyout_Simple.xaml
@@ -1,0 +1,54 @@
+<UserControl
+	x:Class="Uno.UI.Samples.Content.UITests.MenuFlyout.MenuFlyout_Simple"
+	xmlns:controls="using:Uno.UI.Samples.Controls"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:local="using:Uno.UI.Samples.Content.UITests"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	xmlns:u="using:Uno.UI.Samples.Controls"
+	xmlns:uBehaviors="using:Uno.UI.Samples.Behaviors"
+	xmlns:uConverters="using:Uno.UI.Samples.Converters"
+	xmlns:ios="http://uno.ui/ios"
+	xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:android="http://uno.ui/android"
+	mc:Ignorable="d ios android"
+	d:DesignHeight="2000"
+	d:DesignWidth="400">
+
+	<UserControl.Resources>
+		<uConverters:FromNullableBoolToVisibilityConverter x:Key="boolToVisibility" />
+	</UserControl.Resources>
+
+	<controls:SampleControl SampleDescription="Description for sample of MenuFlyout_Simple">
+		<controls:SampleControl.SampleContent>
+			<DataTemplate>
+				<StackPanel>
+					<CheckBox IsChecked="{Binding [IsOption2Visible], Mode=TwoWay}"
+										Content="Is option 2 visible?" />
+
+					<Button Content="Show Options"
+									Height="50">
+						<Button.Flyout>
+							<MenuFlyout>
+								<MenuFlyoutItem Text="Option 1"
+																CommandParameter="Option 1"
+																Command="{Binding [SelectOption]}" />
+								
+								<MenuFlyoutItem Text="Option 2"
+																CommandParameter="Option 2"
+																Command="{Binding [SelectOption]}"
+																Visibility="{Binding [IsOption2Visible], Converter={StaticResource boolToVisibility}}"/>
+
+								<MenuFlyoutItem Text="Option 3"
+																CommandParameter="Option 3"
+																Command="{Binding [SelectOption]}" />
+							</MenuFlyout>
+						</Button.Flyout>
+					</Button>
+					<TextBlock Text="{Binding [SelectedOption]}" FontSize="20" />
+				</StackPanel>
+			</DataTemplate>
+		</controls:SampleControl.SampleContent>
+	</controls:SampleControl>
+</UserControl>
