+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Flyout/MenuFlyout_Simple.xaml.cs
@@ -1,0 +1,14 @@
+using Uno.UI.Samples.Controls;
+using Windows.UI.Xaml.Controls;
+
+namespace Uno.UI.Samples.Content.UITests.MenuFlyout
+{
+	[SampleControlInfo("Flyout", "MenuFlyout_Simple", typeof(MenuFlyoutViewModel))]
+	public sealed partial class MenuFlyout_Simple : UserControl
+	{
+		public MenuFlyout_Simple()
+		{
+			this.InitializeComponent();
+		}
+	}
+}
