--- src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/FlyoutTests/Flyout_Simple.xaml.cs
@@ -1,28 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Uno.UI.Samples.Controls;
-
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
-
-namespace Uno.UI.Samples.Content.UITests.Flyout
-{
-	[SampleControlInfoAttribute("Flyout", "Flyout_Simple")]
-	public sealed partial class Flyout_Simple : UserControl
-	{
-		public Flyout_Simple()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
