--- src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/FlyoutTests/Flyout_Vanilla.xaml.cs
@@ -1,38 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Uno.UI.Samples.Controls;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
-
-// The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
-
-namespace UITests.Shared.Windows_UI_Xaml_Controls.FlyoutTests
-{
-	[SampleControlInfo("Flyout", "Vanilla")]
-	public sealed partial class Flyout_Vanilla : UserControl
-    {
-        public Flyout_Vanilla()
-        {
-            this.InitializeComponent();
-		}
-
-		private void DeleteConfirmation_Click(object sender, RoutedEventArgs e)
-		{
-			Flyout f = this.Control1.Flyout as Flyout;
-			if (f != null)
-			{
-				f.Hide();
-			}
-		}
-	}
-}
